refactor: renames root slice for clarity
